# Update Notes
## 更新数据类

* 用`copy on write`机制重写`NoteGridData`, `Page`, `Beat`
* 修改`NoteGridData`代替`NoteDataCode`作为`DataComponent`
* 没有修改底层数据结构，不会影响现有数据的使用
* 规范部分变量名：`pageNum`, `beatNum`
* 将部分形参类型从`byte`改为`int`
* 改进配套的工具类
* 修改现有的 Player, GUI 以适应新的数据类

## 改进八音盒

* 现在八音盒的叠加功能不再修改原有数据并将其作为结果吐出，而是生成一个新纸带存放叠加后的数据

## TODO

* 在修改现有 GUI 时，发现了很多设计得不够好的地方和一些小 bug，后面得重写一下

+++ src/main/java/io/github/c20c01/cc_mb/player/MusicBoxPlayer.java
@@ -44,7 +44,7 @@ public class MusicBoxPlayer implements NoteGridIteratorListener {
     }
 
     @Override
-    public boolean onBeat(Beat beat, byte beatNumber) {
+    public boolean onBeat(Beat beat, int beatNumber) {
         listener.onBeat(beat, beatNumber); // always return true
 
         if (beat.isEmpty() || config.isSilent()) return true;
@@ -57,7 +57,7 @@ public class MusicBoxPlayer implements NoteGridIteratorListener {
     }
 
     @Override
-    public void onPageChanged(byte pageNum) {
+    public void onPageChanged(int pageNum) {
         listener.onPageChanged(pageNum);
     }
